validation setup and dto config
