+++ api/src/global.dto.ts
@@ -1,0 +1,7 @@
+export class ListAllEntities {
+    where?: string
+    skip?: number
+    take?: number
+    cursor?: number
+    orderBy?: string
+}
