+++ api/src/main.ts
@@ -1,6 +1,7 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { config } from 'dotenv';
+import { ValidationPipe } from '@nestjs/common';
 declare const module: any;
 
 config();
@@ -8,6 +9,7 @@ const PORT = process.env.PORT;
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
+  app.useGlobalPipes(new ValidationPipe);
   await app.listen(PORT);
 
   console.log(`API Online. Access http://localhost:${PORT}`)
